+++ src/Searchable.php
@@ -113,9 +113,10 @@ trait Searchable
     /**
      * Make all instances of the model searchable.
      *
+     * @param  int  $chunk
      * @return void
      */
-    public static function makeAllSearchable()
+    public static function makeAllSearchable($chunk = null)
     {
         $self = new static;
 
@@ -126,7 +127,7 @@ trait Searchable
                 $query->withTrashed();
             })
             ->orderBy($self->getKeyName())
-            ->searchable();
+            ->searchable($chunk);
     }
 
     /**