+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/scout/compare/v8.0.0...master)
+## [Unreleased](https://github.com/laravel/scout/compare/v8.2.0...master)
+
+
+## [v8.2.0 (2020-08-04)](https://github.com/laravel/scout/compare/v8.0.1...v8.2.0)
+
+### Added
+- Identifying users ([#411](https://github.com/laravel/scout/pull/411), [fe28ab2](https://github.com/laravel/scout/commit/fe28ab26bf1e5c9c3b46f2535bea746b69fa6fb1))
+
+
+## [v8.1.0 (2020-07-14)](https://github.com/laravel/scout/compare/v8.0.1...v8.1.0)
+
+### Added
+- Optional param for chunk size on `scout:import` ([#407](https://github.com/laravel/scout/pull/407))
+
+
+## [v8.0.1 (2020-04-21)](https://github.com/laravel/scout/compare/v8.0.0...v8.0.1)
+
+### Fixed
+- Merge default scout configs ([#402](https://github.com/laravel/scout/pull/402))
 
 
 ## [v8.0.0 (2020-03-03)](https://github.com/laravel/scout/compare/v7.2.1...v8.0.0)